+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove the document_embed function.
 - Support MongoDB outputs query
 - Make "create a table" compulsory
+- All datatypes should be wrapped with a Schema
 
 #### Bug Fixes